Fix location in the Confirm Import Settings dialog
Restore the prior patch to fix the location in the Confirm Import
Settings dialog. Normally, NetBeans looks in '$HOME/.netbeans' for
settings created by a previous version of the IDE on Linux. The patch
changes this location to '$SNAP_USER_DATA'.

Add the NetBeans version to the user and cache directories:

  --userdir "$SNAP_USER_DATA/$SNAP_VERSION"
  --cachedir "$SNAP_USER_COMMON/$SNAP_VERSION"