Remove patch and add proxy networking properties
Patching the NetBeans startup script fails silently for remote builds
on Launchpad, whether it occurs in the part's build directory or
after the build output is moved to the install directory.

Remove the patch so that the NetBeans Snap package behaves just as
the NetBeans Binaries package from Apache, storing its data files
under the '~/.netbeans' directory and its cache files under the
'~/.cache/netbeans' directory.

Parse the proxy environment variables, if defined, and add the
corresponding Java networking properties to the 'ANT_OPTS' environment
variable.