Upgrade to latest NetBeans 13 release candidate
Build from the NetBeans 13 release branch on GitHub to get the most
recent release candidate. Also upgrade to the 'core20' base.

The 'override-build' scriptlet quits after building NetBeans with
Apache Ant version 1.10.7 in the 'core20' base. The scriptlet
completes when using either version 1.10.5 in the 'core18' base or
version 1.10.12 of the Apache Ant Snap package. Switch to the Snap
package to work around the problem.

Scriptlets that override plugins run with the shells and options below:

  Base      Shell       Options
  core18    /bin/sh     set -xe
  core20    /bin/bash   set -xeuo pipefail

See:

  https://github.com/snapcore/snapcraft/
    blob/main/snapcraft/internal/pluginhandler/__init__.py

Initialize 'proxies' to avoid errors with the Bash 'nounset'
option, and add read permissions for group and other to the file
'etc/netbeans.conf' so that NetBeans can start.