Change default local repository in global settings
Modify the Apache Maven global settings file so that the local
repository is in a location accessible to this snap package. The
original default of '${user.home}/.m2/repository' is in a hidden
directory inaccessible to strictly-confined snaps.

Remove the 'netbeans' and 'maven' subdirectories of '$SNAP_USER_COMMON'
and place their contents directly under the parent directory.