Override value of 'netbeans.default_userdir_root'
Even when the '--userdir' option is set, its parent directory is
determined independently by the 'netbeans.default_userdir_root' system
property. NetBeans uses the property to find the Maven local repository
for indexing and to find any prior version's settings for importing.

Override the property's default value of '~/.netbeans' by redefining
it to '$HOME' in the JVM options of the NetBeans launcher. Note that
HOME is rewritten by 'snapd' to '~/snap/strictly-netbeans/current'.